chore: move tests to _whitebox_test file
